+++ Source/FileWatcherExTests/scenario/create_file.csv
@@ -1,2 +1,2 @@
-FullPath,EventName,OldFullPath,DiffInTicks,DiffInMilliseconds
-C:\temp\fwtest\a.txt,created,,0,0
+Directory,FileName,EventName,OldFileName,DiffInTicks,DiffInMilliseconds
+C:\temp\fwtest,a.txt,created,,0,0
